requirements.txt
flask
pinecone-client
openai
python-dotenv
pymysql
requests
beautifulsoup4
lxml
tiktoken
numpy